services/backend/requirements.txt
# FastAPI
fastapi==0.118.0
uvicorn==0.37.0
pydantic==2.11.9
pydantic[email]==2.11.9
python-dotenv==1.1.1
PyJWT==2.10.1
sqlmodel==0.0.25
sqlalchemy==2.0.43
alembic==1.16.5
psycopg2-binary==2.9.10

# External API
openai==2.0.1
workos==5.28.0